Har lagt til div funksjonaliteer sjekk trello

+++ Arduino/ArduinoMega_main/main/actuator.cpp
@@ -54,7 +54,7 @@ void Actuator::computePID(){
 
 
 //Sets a new setpoint
-void Actuator::setSetpoint(int r){
+void Actuator::setSetpoint(float r){
         _setPoint = r;
 }
